chore(fixes): eslint
- update eslint
- changed components extensions to jsx
- correct files with eslint errors

[finishes #151154501]
