+++ .eslintrc
@@ -1,94 +1,39 @@
 {
   "root": true,
   "extends": "airbnb",
-  "plugins": [
-    "react"
-  ],
-  "parserOptions": {
-    "ecmaVersion": 6,
-    "sourceType": "module",
-    "ecmaFeatures": {
-      "jsx": true
-    }
-  },
   "env": {
     "node": true,
     "es6": true,
     "mocha": true,
-    "browser": true,
-    "jquery": true
-  },
-  "globals":{
-    "Materialize": true
+    "jasmine": true
   },
   "rules": {
-    "max-len": [1, 80, 2],
-    "indent": [ 1, 2 ],
     "one-var": 0,
     "one-var-declaration-per-line": 0,
     "new-cap": 0,
     "consistent-return": 0,
     "no-param-reassign": 0,
+    "import/no-unresolved": 0,
     "comma-dangle": 0,
     "curly": ["error", "multi-line"],
-    "import/no-unresolved": [2, { "commonjs": true }],
-    "no-shadow": ["error", { "allow": ["req", "res", "err"] }],
+    "no-shadow": ["error", { "allow": ["req", "done", "res", "err"] }],
     "valid-jsdoc": ["error", {
       "requireReturn": true,
       "requireReturnType": true,
       "requireParamDescription": false,
-      "requireReturnDescription": true,
-      "FunctionDeclaration": true,
-      "MethodDefinition": false,
-      "ClassDeclaration": false
+      "requireReturnDescription": true
     }],
-    "quotes": 0,
-    "no-console": 1,
-    "no-debugger": 1,
-    "no-var": 1,
-    "semi": [1, "always"],
-    "no-trailing-spaces": 0,
-    "eol-last": 0,
-    "no-unused-vars": 0,
-    "no-underscore-dangle": 0,
-    "no-alert": 0,
-    "no-lone-blocks": 0,
-    "jsx-quotes": 1,
-    "react/display-name": [ 1, {"ignoreTranspilerName": false }],
-    "react/forbid-prop-types": [1, {"forbid": ["any"]}],
-    "react/jsx-boolean-value": 1,
-    "react/jsx-closing-bracket-location": 0,
-    "react/jsx-curly-spacing": 1,
-    "react/jsx-indent-props": 0,
-    "react/jsx-key": 1,
-    "react/jsx-max-props-per-line": 0,
-    "react/jsx-no-bind": 1,
-    "react/jsx-no-duplicate-props": 1,
-    "react/jsx-no-literals": 0,
-    "react/jsx-no-undef": 1,
-    "react/jsx-pascal-case": 1,
-    "react/jsx-sort-prop-types": 0,
-    "react/jsx-sort-props": 0,
-    "react/jsx-uses-react": 1,
-    "react/jsx-uses-vars": 1,
-    "react/no-danger": 1,
-    "react/no-did-mount-set-state": 1,
-    "react/no-did-update-set-state": 1,
-    "react/no-direct-mutation-state": 1,
-    "react/no-multi-comp": 1,
-    "react/no-set-state": 0,
-    "react/no-unknown-property": 1,
-    "react/prefer-es6-class": 1,
-    "react/prop-types": 1,
-    "react/react-in-jsx-scope": 1,
-    "react/self-closing-comp": 1,
-    "react/sort-comp": 1
-    // "require-jsdoc": ["error", {
-    //     "require": {
-    //         "FunctionDeclaration": true,
-    //         "MethodDefinition": true,
-    //         "ClassDeclaration": true
-    //     }
-    // }]
+    "require-jsdoc": ["error", {
+        "require": {
+            "FunctionDeclaration": true,
+            "MethodDefinition": true,
+            "ClassDeclaration": true
+        }
+    }]
+  },
+  "globals": {
+    "localStorage": true,
+    "window": true,
+    "document": true
   }
 }
