+++ client/src/constants.js
@@ -14,3 +14,4 @@ export const RESET_PASSWORD = 'RESET_PASSWORD';
 export const LOAD_UNREAD_MESSAGES = 'LOAD_UNREAD_MESSAGES';
 export const LOAD_READ_MESSAGES = 'LOAD_READ_MESSAGES';
 export const LOAD_COUNT = 'LOAD_COUNT';
+export const LOAD_PAGE_COUNT = 'LOAD_PAGE_COUNT';
