+++ client/src/index.js
@@ -1,14 +1,13 @@
 import React from 'react';
+import 'react-materialize';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { BrowserRouter as Router} from 'react-router-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
 import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
-import App from './components/App';
+import App from './components/App.jsx';
 import rootReducer from './reducers/rootReducer';
-import 'react-materialize';
 import { authenticateUser } from './components/auth';
-import { RELOAD_USER_IN } from './constants';
 
 const store = createStore(
   rootReducer,
