+++ package.json
@@ -59,10 +59,11 @@
     "dotenv": "^4.0.0",
     "enzyme": "^2.9.1",
     "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^15.1.0",
     "eslint-config-airbnb-base": "^11.2.0",
     "eslint-plugin-flowtype": "^2.35.0",
     "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
     "eslint-plugin-react": "^7.1.0",
     "eslint-watch": "^3.1.2",
     "eventsource-polyfill": "^0.9.6",
@@ -95,6 +96,7 @@
     "react-dom": "^15.6.1",
     "react-materialize": "^1.0.6",
     "react-modal": "^2.3.2",
+    "react-paginate": "^4.4.4",
     "react-redux": "^5.0.5",
     "react-router": "^4.1.2",
     "react-router-dom": "^4.1.2",
