+++ webpack.config.dev.js
@@ -1,10 +1,10 @@
-import webpack from 'webpack';
-import path from 'path';
+const webpack = require('webpack');
+const path = require('path');
 
-export default {
-  debug: true,
-  devtool: 'cheap-module-eval-source-map',
-  noInfo: false,  // to change to true soon
+module.exports = {
+  // debug: true,
+  devtool: 'inline-source-map',
+  // noInfo: false,  // to change to true soon
   entry: [
     'eventsource-polyfill',
     'webpack-hot-middleware/client?reload=true',
@@ -12,16 +12,20 @@ export default {
   ],
   target: 'web',
   output: {
-    path: __dirname + '/client/dist', // Note: Physical files are only output by the production build task `npm run build`.
+    path: path.join(__dirname, '/client/dist/js'), // Note: Physical files are only output by the production build task `npm run build`.
     publicPath: '/',
     filename: 'bundle.js'
   },
-  devServer: {
-    contentBase: './client/src'
-  },
+  // devServer: {
+  //   contentBase: './client/src'
+  // },
   plugins: [
     new webpack.HotModuleReplacementPlugin(),
-    new webpack.NoErrorsPlugin()
+    new webpack.LoaderOptionsPlugin({
+      debug: true
+    }),
+    new webpack.NoEmitOnErrorsPlugin()
+    // new webpack.NoErrorsPlugin()
   ],
   node: {
     net: 'empty',
@@ -31,18 +35,27 @@ export default {
   },
   module: {
     loaders: [
-      {test: /\.js$/, include: path.join(__dirname, './client/src'), loaders: ['babel']},
-      {test: /(\.css)$/, loaders: ['style', 'css']},
-      {test: /\.eot(\?v=\d+\.\d+\.\d+)?$/, loader: "file"},
-      {test: /\.(woff|woff2)$/, loader: "url?prefix=font/&limit=5000"},
-      {test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/, loader: "url?limit=10000&mimetype=application/octet-stream"},
-      {test: /\.svg(\?v=\d+\.\d+\.\d+)?$/, loader: "url?limit=10000&mimetype=image/svg+xml"},
-      {test: /\.(jpe?g|png|gif|svg)$/i,
+      {
+        test: /.jsx?$/,
+        include: [path.join(__dirname, './client/src')],
+        exclude: /(node_modules|bower_components)/,
+        loaders: ['babel-loader']
+      },
+      { test: /(\.css)$/, loaders: ['style-loader', 'css-loader'] },
+      { test: /\.eot(\?v=\d+\.\d+\.\d+)?$/, loader: 'file-loader' },
+      { test: /\.(woff|woff2)$/, loader: 'url-loader?prefix=font/&limit=5000' },
+      { test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/, loader: 'url-loader?limit=10000&mimetype=application/octet-stream' },
+      { test: /\.svg(\?v=\d+\.\d+\.\d+)?$/, loader: 'url-loader?limit=10000&mimetype=image/svg+xml' },
+      {
+        test: /\.(jpe?g|png|gif|svg)$/i,
         loaders: [
-            'file-loader?hash=sha512&digest=hex&name=[hash].[ext]',
-            'image-webpack-loader?bypassOnDebug&optimizationLevel=7&interlaced=false'
+          'file-loader?hash=sha512&digest=hex&name=[hash].[ext]',
+          'image-webpack-loader?bypassOnDebug&optimizationLevel=7&interlaced=false'
         ]
       }
     ]
+  },
+  resolve: {
+    extensions: ['.js', '.jsx']
   }
 };
